Make the sum of squares column a bigint
git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.learning_network/trunk@63446 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/nti/learning_network/database/meta_mixins.py
@@ -10,6 +10,7 @@ logger = __import__('logging').getLogger(__name__)
 
 from sqlalchemy import Column
 from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy import ForeignKey
 from sqlalchemy import DateTime
 
@@ -40,4 +41,4 @@ class StatMixin( object ):
 
 	total_duration = Column('total_duration', Integer, nullable=False, autoincrement=False, default=0 )
 	count = Column('count', Integer, nullable=False, autoincrement=False, default=0 )
-	sum_of_squares = Column('sum_of_squares', Integer, nullable=False, autoincrement=False, default=0 )
+	sum_of_squares = Column('sum_of_squares', BigInteger, nullable=False, autoincrement=False, default=0 )
